updated database interface for supervised schema refinement

+++ src/database.py
@@ -6,13 +6,44 @@ import sqlalchemy
 import sqlalchemy_schemadisplay
 import snowflake.connector
 import networkx as nx
+from typing import List, Dict
 import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 from src.database_utils import get_view_name_from_definition
 
+"""
+NLQuery class
+"""
+class NLQuery:
+    def __init__(self, nl_query: str, sql_query: str, database_name: str, evidence: str = None):
+        self._nl_query = nl_query
+        self._sql = sql_query
+        self._database = database_name
+        self._evidence = evidence
+
+    def get_nl_query(self):
+        return self._nl_query
+
+    def get_sql_query(self):
+        return self._sql
+
+    def get_database(self):
+        return self._database
+
+    def get_evidence(self):
+        return self._evidence
+
+    def __str__(self):
+        evidence = f"\nEvidence : {self._evidence}" if self._evidence else ""
+        return f"Query    : {self._nl_query}" + "\n" + f"SQL      : {self._sql} " + evidence
+
+
+"""
+Database class
+"""
 class Database:
-    def __init__(self, database_name):
+    def __init__(self, database_name: str):
         self.db_name = database_name
 
     @property
@@ -25,74 +56,102 @@ class Database:
         """
         raise NotImplementedError
     
-    def get_columns_of_table(self, table_name):
+    def get_columns_of_table(self, table_name: str):
         """
         Get a list of column names for a given table.
         """
         raise NotImplementedError
+
+    def run_sql_query(self, query: str):
+        """
+        Run a SQL query on the database.
+        """
+        raise NotImplementedError
     
-    def schema_dictionary(self, include_views=False):
+    def schema_dictionary(self, include_views: bool = False):
         """
         Get schema of the database in dictionary format.
         Table name is the key and the value is a list of column names.
         """
         raise NotImplementedError
     
-    def schema_wording(self, selected_tables=None, include_sample_data=True, sample_size=5):
+    def schema_wording(self, selected_tables: List[str] = None, include_sample_data: bool = True, sample_size: int = 5):
         """
         Generate a textual description of the schema of the database.
         """
         raise NotImplementedError
 
-    def schema_graph(self, save_dir=None):
+    def schema_graph(self, save_dir: str = None):
         """
         Generate a networkx graph object representing the schema of the database.
         """
         raise NotImplementedError
     
-    def materialize_view(self, view_definition, verbose=True, persist=False):
+    def materialize_view(self, view_definition: str, verbose: bool = True, persist: bool = False):
         """
         Materialize a view in the database.
         """
         raise NotImplementedError
 
+
 """
 SQLite Database class
 """
 class SQLiteDatabase(Database):
-    def __init__(self, database_name, database_dir):
+    def __init__(self, database_name: str, database_dir: str, query_log_full_path: str = None):
         super().__init__(database_name)
-        self.db_dir = database_dir
+        self._db_dir = database_dir
+        self._query_log_full_path = query_log_full_path
 
     def get_tables(self):
         """
         Get a list of table names in the SQLite database.
         """
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
         cursor.execute("SELECT name FROM sqlite_master WHERE type = 'table'")
         tables = [table[0] for table in cursor.fetchall()]
         conn.close()
         return tables
     
-    def get_columns_of_table(self, table_name):
+    def get_columns_of_table(self, table_name: str):
         """
         Get a list of column names for a given table in the SQLite database.
         """
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
         cursor.execute(f"PRAGMA table_info({table_name})")
         columns = [column[1] for column in cursor.fetchall()]
         conn.close()
         return columns
+
+    def get_nl_queries(self):
+        """
+        Get a list of natural language queries from the query log.
+        """
+        if not self._query_log_full_path:
+            return []
+        try:
+            with open(self._query_log_full_path, "r") as f:
+                lines = json.load(f)
+            queries = []
+            for q in lines:
+                if q['db_id'] == self.database_name:
+                    if 'evidence' in q:
+                        queries.append(NLQuery(q['question'], q['SQL'], q['db_id'], q['evidence']))
+                    else:
+                        queries.append(NLQuery(q['question'], q['SQL'], q['db_id']))
+        except Exception as e:
+            raise Exception(f"Error in reading query log: {e}")
+        return queries
     
-    def schema_dictionary(self, include_views=False):
+    def schema_dictionary(self, include_views: bool = False):
         """
         Get schema of the SQLite database in dictionary format. Table (and view) name is the key and the list of column names is the value.
         include_views: If True, include view names as keys in the schema dictionary.
         """
         schema = {}
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
 
         # fetch table names
@@ -111,7 +170,6 @@ class SQLiteDatabase(Database):
                 cursor.execute("PRAGMA table_info({})".format(table))
                 schema[table] = [str(col[1].lower()) for col in cursor.fetchall()]
             except Exception as e:
-                # print(f"Error in table {table}. {e}")
                 pass
         
         # Close the connection
@@ -119,12 +177,12 @@ class SQLiteDatabase(Database):
 
         return schema
     
-    def schema_wording(self, selected_tables=None, include_sample_data=True, sample_size=5):
+    def schema_wording(self, selected_tables: List[str] = None, include_sample_data: bool = True, sample_size: int = 5):
         """
         Generate a textual description of the schema of the SQLite database, in the form of Data Definition Language (DDL) statements.
         """
         # Connect to the SQLite database
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
 
         # Query to get the tables
@@ -182,12 +240,12 @@ class SQLiteDatabase(Database):
 
         return DDL
 
-    def schema_wording_simple(self, selected_tables=None, include_sample_data=True, sample_size=5):
+    def schema_wording_simple(self, selected_tables: List[str] = None, include_sample_data: bool = True, sample_size: int = 5):
         """
         Generate a textual description of the schema of the SQLite database.
         """
         # Connect to the SQLite database
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
 
         # Query to get the tables
@@ -244,7 +302,7 @@ class SQLiteDatabase(Database):
         Generate a networkx graph object representing the schema of the SQLite database.
         """
         # Connect to the SQLite database
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
 
         # Query to get the tables
@@ -277,12 +335,12 @@ class SQLiteDatabase(Database):
         
         return G
     
-    def materialize_view(self, view_definition, verbose=True, persist=False):
+    def materialize_view(self, view_definition: str, verbose: bool = True, replace: bool = True, persist: bool = False):
         """
         Materialize a view in the SQLite database.
         """
         # Connect to the database
-        conn = sqlite3.connect(self.db_dir)
+        conn = sqlite3.connect(self._db_dir)
         cursor = conn.cursor()
         
         # Get the view name 
@@ -301,10 +359,13 @@ class SQLiteDatabase(Database):
             except Exception as e:
                 # If the error says the view already exists, drop the view and try again
                 if "already exists" in str(e):
-                    if verbose:
-                        print(f"View {view_name} already exists. Dropping the view and running the query again...")
-                    cursor.execute(f"DROP VIEW {view_name}")
-                    conn.commit()
+                    if not replace:
+                        return f"Error in creating view {view_name}. View already exists."
+                    else:
+                        if verbose:
+                            print(f"View {view_name} already exists. Dropping the view and running the query again...")
+                        cursor.execute(f"DROP VIEW {view_name}")
+                        conn.commit()
                 # Else return the error
                 else:
                     if verbose:
@@ -330,12 +391,27 @@ class SQLiteDatabase(Database):
 
         return f"View {view_name} successfully defined."
 
+    def run_sql_query(self, query: str):
+        """
+        Run a SQL query on the SQLite database.
+        """
+        try:
+            conn = sqlite3.connect(self._db_dir)
+            cursor = conn.cursor()
+            cursor.execute(query)
+            results = cursor.fetchall()
+            conn.close()
+        except Exception as e:
+            conn.close()
+            return f"Error in executing query: {e}"
+        return results
+
 
 """
 Snowflake Database class
 """
 class SnowflakeDatabase(Database):
-    def __init__(self, database_name, snowflake_config_file):
+    def __init__(self, database_name: str, snowflake_config_file: str):
         super().__init__(database_name)
         try:
             self.snowflake_config = json.load(open(snowflake_config_file))[database_name]
@@ -349,23 +425,25 @@ class SnowflakeDatabase(Database):
         assert 'database' in self.snowflake_config, "Snowflake database not found in the config file."
         assert 'schema' in self.snowflake_config, "Snowflake schema not found in the config file"
 
-    def open_connection(self):
+    def _open_connection(self):
         """
         Open a connection to the Snowflake database.
         """
-        # Connect to Snowflake
-        ctx = snowflake.connector.connect(
-            user=self.snowflake_config['user'],
-            password=self.snowflake_config['password'],
-            account=self.snowflake_config['account'],
-            role=self.snowflake_config['role'],
-        )
-
-        # Connect to the database
-        cs = ctx.cursor()
-        cs.execute(f"USE WAREHOUSE {self.snowflake_config['warehouse']}")
-        cs.execute(f"USE DATABASE {self.snowflake_config['database']}")
-        cs.execute(f"USE SCHEMA {self.snowflake_config['schema']}")
+        try:
+            # Connect to Snowflake
+            ctx = snowflake.connector.connect(
+                user=self.snowflake_config['user'],
+                password=self.snowflake_config['password'],
+                account=self.snowflake_config['account'],
+                role=self.snowflake_config['role'],
+            )
+            # Connect to the database
+            cs = ctx.cursor()
+            cs.execute(f"USE WAREHOUSE {self.snowflake_config['warehouse']}")
+            cs.execute(f"USE DATABASE {self.snowflake_config['database']}")
+            cs.execute(f"USE SCHEMA {self.snowflake_config['schema']}")
+        except Exception as e:
+            raise Exception(f"Error in connecting to Snowflake database: {e}")
 
         return ctx, cs
 
@@ -374,29 +452,37 @@ class SnowflakeDatabase(Database):
         Get a list of table names in the Snowflake database.
         """
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
 
         # fetch table names
         cs.execute("SHOW TABLES")
         tables = [table[1] for table in cs.fetchall()]
 
+        # Close the connection
+        cs.close()
+        ctx.close()
+
         return tables
     
-    def get_columns_of_table(self, table_name):
+    def get_columns_of_table(self, table_name: str):
         """
         Get a list of column names for a given table in the Snowflake database.
         """
 
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
 
         # fetch table info
         cs.execute(f"DESCRIBE TABLE {table_name}")
         columns = [column[0] for column in cs.fetchall()]
 
+        # Close the connection
+        cs.close()
+        ctx.close()
+
         return columns
     
-    def schema_dictionary(self, include_views=False):
+    def schema_dictionary(self, include_views: bool = False):
         """
         Get schema of the Snowflake database in dictionary format. Table (and view) name is the key and the list of column names is the value.
         include_views: If True, include view names as keys in the schema dictionary.
@@ -404,7 +490,7 @@ class SnowflakeDatabase(Database):
         schema = {}
     
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
 
         # fetch table names
         cs.execute("SHOW TABLES")
@@ -427,12 +513,12 @@ class SnowflakeDatabase(Database):
 
         return schema
     
-    def schema_wording(self, selected_tables=None, include_sample_data=True, sample_size=5):
+    def schema_wording(self, selected_tables: List[str] = None, include_sample_data: bool = True, sample_size: int = 5):
         """
         Generate a textual description of the schema of the Snowflake database, in the form of Data Definition Language (DDL) statements.
         """
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
 
         # Query to get the tables
         tables = self.get_tables()
@@ -487,12 +573,12 @@ class SnowflakeDatabase(Database):
 
         return DDL
 
-    def schema_wording_simple(self, selected_tables=None, include_sample_data=True, sample_size=5):
+    def schema_wording_simple(self, selected_tables: List[str] = None, include_sample_data: bool = True, sample_size: int = 5):
         """
         Generate a textual description of the schema of the Snowflake database.
         """
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
         
         # Query to get the tables
         tables = self.get_tables()
@@ -549,7 +635,7 @@ class SnowflakeDatabase(Database):
         Generate a networkx graph object representing the schema of the Snowflake database.
         """
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
 
         # Query to get the tables
         tables = self.get_tables()
@@ -589,13 +675,13 @@ class SnowflakeDatabase(Database):
         
         return G
     
-    def materialize_view(self, view_definition, verbose=True, persist=False):
+    def materialize_view(self, view_definition: str, verbose: bool = True, replace: bool = True, persist: bool = False):
         """
         Materialize a view in the Snowflake database.
         """
         
         # Get connection and cursor
-        ctx, cs = self.open_connection()
+        ctx, cs = self._open_connection()
         
         # Get the view name
         view_name = get_view_name_from_definition(view_definition)
@@ -611,9 +697,12 @@ class SnowflakeDatabase(Database):
             except Exception as e:
                 # If the error says the view already exists, drop the view and try again
                 if "already exists" in str(e):
-                    if verbose:
-                        print(f"View {view_name} already exists. Dropping the view and running the query again...")
-                    cs.execute(f"DROP VIEW {view_name}")
+                    if not replace:
+                        return f"Error in creating view {view_name}. View already exists."
+                    else:
+                        if verbose:
+                            print(f"View {view_name} already exists. Dropping the view and running the query again...")
+                        cs.execute(f"DROP VIEW {view_name}")
                 # Else return the error
                 else:
                     if verbose:
@@ -635,3 +724,19 @@ class SnowflakeDatabase(Database):
         ctx.close()
 
         return f"View {view_name} successfully defined."
+
+    def run_sql_query(self, query: str):
+        """
+        Run a SQL query on the Snowflake database.
+        """
+        ctx, cs = self._open_connection()
+        try:
+            cs.execute(query)
+            results = cs.fetchall()
+            cs.close()
+            ctx.close()
+        except Exception as e:
+            cs.close()
+            ctx.close()
+            return f"Error in executing query: {e}"
+        return results
